interview_orb/requirements.txt
google-cloud-speech>=2.21.0
google-cloud-texttospeech>=2.16.0
google-genai>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pydub>=0.25.1
pydantic>=2.0.0
python-dotenv>=1.0.0
